feat: added create option for cottages

+++ isa-project/src/main/java/isaproject/repository/CottageImageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import isaproject.model.CottageImage;
 
-public interface CottageImageRepository extends JpaRepository<CottageImage, Integer> {
+public interface CottageImageRepository extends JpaRepository<CottageImage, Long> {
 
 }
