+++ isa-project/src/main/java/isaproject/repository/CottageQuickReservationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import isaproject.model.CottageQuickReservation;
 
-public interface CottageQuickReservationRepository extends JpaRepository<CottageQuickReservation, Integer>{
+public interface CottageQuickReservationRepository extends JpaRepository<CottageQuickReservation, Long>{
 
 }
